pyproject.toml
[tool.poetry]
name = "testframework-pytest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.0.1"
pytest-check = "^2.3.1"
requests = "^2.31.0"
selenium = "^4.17.2"
isort = "^5.9.3"
webdriver-manager = "^4.0.1"
allure-pytest = "^2.13.2"
pillow = "10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"